ml/requirements.txt
flask
flask-cors
google-generativeai
python-dotenv
pymupdf
numpy
pandas
scikit-learn